travelhub-experiment-2/autorizador/requirements.txt
flask==3.0.3
PyJWT==2.8.0
cryptography>=41.0.0
psycopg2-binary==2.9.10
